+++ So1omon/src/main/java/com/kh/so1omon/board/model/service/BoardServiceImp.java
@@ -73,12 +73,12 @@ public class BoardServiceImp implements BoardService{
 	}
 
 	@Override
-	public ArrayList<Board> selectAllBoardListAD(int userNo) {
+	public ArrayList<Board> selectAllBoardListAD(long userNo) {
 		return bDao.selectAllBoardListAD(sqlSession, userNo);
 	}
 
 	@Override
-	public ArrayList<Reply> selectReplyListAD(int userNo) {
+	public ArrayList<Reply> selectReplyListAD(long userNo) {
 		return bDao.selectReplyListAD(sqlSession, userNo);
 	}
